Remove support for Ubuntu Xenial (16)
The age of some of the packages available now create problems for some Python
package installs. Specifically the available pip version is 8.1.1, and the
available Python versions are 2.7.11 and 3.5.1 respectively. The problem is
that pip < 9 is not PEP 503 (https://www.python.org/dev/peps/pep-0503/)
compliant. This means they do not respect package directives for Python
versionsy such as `python_requires`. This results in a problem for two
reasons:

1. Xenial only has Py3.5 and our own Python projects are built for Py3.6+
2. Python packages in general are dropping support for <Py3.6
3. Without PEP503 compliance, the pip version on Xenial hosts will install
   package versions that are incompatible with the Python3 package version
   it has available.

Since Xenial standard support ends 2021/04 (just over two months from this
commit), and is EoL in 2024/04, it is not worth trying to work around
supporting this platform that we are not actively using.