Use the full path for source Docker images
This ensures that when a Docker image is pulled for molecule's testing
the expected source image is used regardless of what registry is
configured as the default on the host system. This will also make it
trivial to convert to using the GitHub Container Registry or any other
Open Container Initiative (OCI) container registry for source images.