Restore externally managed Python environment indicator
In SystemD-enabled Docker images sources from geerlingguy he has made
the decision to remove the indicator for an externally managed Python
environment if the host is configured with one. This impacts Debian
Bookworm and Trixie, as well as Ubuntu 24.04 (Noble Numbat), and as far
as we know any future versions of these distributions. This will help
ensure that our Ansible role testing is done against images that
reflect how a host will be configured when used to create AMIs.