Install our Python linting tools
This is useful because our Molecule test suites are all written in
Python, and we use these tools for Python code formatting.  Including
these tools in the virtual environment for developers will allow their
editors automatically format Python code.  It also allows developers
to run the tools manually if they wish to do so.

This would be even more useful if pytest-testinfra were types, but
that is not yet the case.  See pytest-dev/pytest-testinfra#619 for
more details about that.