Pin Molecule to <25.2.0
As of Molecule v25.2.0, Molecule no longer adds the current project
directory to the search path for Ansible roles.  This breaks our
current Molecule testing, since Ansible does not find the local copy
of the role being tested.  For now, it makes sense to pin to <25.2.0
until we figure out how to deal with this change.